High Resolution Bot for Reddit
A Reddit bot that finds the highest available resolution of an image. I
am a complete beginner when it comes to using an API. For example, a
user could type '!hiresbot' followed by the image URL. Or, typing
'!hiresbot' could search for a URL in the parent comment (but I'm not
sure how to do that). The bot could return the URL to the highest
resolution image available, or let the user know that there is no higher
resolution available.